REFACTOR: actor
