+++ oceanraft/Cargo.toml
@@ -28,6 +28,8 @@ bytes = { version = "1" }
 prost = { version = "0.11" }
 smallvec = { version = "1" }
 tonic = { version = "0.8", optional = true }
+flume = { version = "0.10.14" }
+
 
 # [dependencies.rocksdb]
 # default-features = false
