+++ oceanraft/src/multiraft/mod.rs
@@ -1,31 +1,22 @@
-mod apply_actor;
 mod config;
 mod error;
 mod multiraft;
-mod multiraft_actor;
 mod proposal;
-// pub mod transport;
 pub mod transport;
-// mod transport_local;
 mod util;
-// mod write;
 mod event;
-mod node;
-mod raft_group;
 mod replica_cache;
 mod response;
 mod rsm;
+mod apply;
+mod group;
+mod msg;
+mod node;
 pub mod storage;
-// pub use transport_local::LocalTransport;
 
 pub use config::Config;
-pub use error::{
-    Error, MultiRaftStorageError, RaftCoreError, RaftGroupError,
-    WriteError,
-};
-pub use event::{
-    ApplyEvent, ApplyMembership, ApplyNoOp, ApplyNormal, CommitEvent, Event, LeaderElectionEvent,
-};
+pub use error::{Error, MultiRaftStorageError, RaftCoreError, RaftGroupError, WriteError};
+pub use event::{Event, LeaderElectionEvent};
 pub use multiraft::{MultiRaft, MultiRaftMessageSender, MultiRaftMessageSenderImpl};
-pub use rsm::StateMachine;
+pub use rsm::{Apply, ApplyMembership, ApplyNoOp, ApplyNormal, StateMachine};
 pub use util::{ManualTick, Ticker};
